requirements.txt
torch==2.4.0
itakello_logging
wandb
transformers
python-dotenv
evaluate
accelerate
weave